+++ src/main/java/subway/application/line/service/FindAllLineService.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import subway.application.line.usecase.FindAllLinesUseCase;
 import subway.domain.line.Line;
 import subway.domain.line.LineRepository;
-import subway.domain.line.LineRoutePath;
+import subway.domain.path.LineRoutePath;
 import subway.ui.dto.response.LineResponse;
 import subway.ui.dto.response.StationResponse;
